transport_species_parsing no longer header only
There is some header dependency hell to be cleaned up after finishing
migrating other headers to implementation.
