+++ src/Makefile.am
@@ -18,6 +18,7 @@ libantioch_la_SOURCES += core/src/chemical_mixture.C
 # parsing
 libantioch_la_SOURCES += parsing/src/ascii_parser.C
 libantioch_la_SOURCES += parsing/src/blottner_parsing.C
+libantioch_la_SOURCES += parsing/src/transport_species_parsing.C
 
 # transport
 libantioch_la_SOURCES += transport/src/transport_mixture.C
